pyproject.toml
[tool.poetry]
name = "tap-asteroids"
version = "0.1.0"
description = ""
authors = ["Ross Perry <perryr16@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
singer-python = "^5.11.0"
requests = "^2.25.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
vcrpy = "^4.1.1"
pytest-vcr = "^1.0.2"
autopep8 = "^1.5.5"
black = {version = "^20.8b1", allow-prereleases = true}
pytest-mock = "^3.5.1"
mock = "^4.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
